src/components: use visualizerSettingsStore
Additionally, move the visualizer settings menu to a separate
component.
