+++ src/components/App/App.tsx
@@ -10,13 +10,8 @@ import * as Tone from "tone";
 import { PreferencesContext, defaultPreferences } from "../../contexts";
 import { NavBar } from "../navbar";
 import { Background } from "../background";
-import {
-  type VisualizerType,
-  type VisualizerColorSource,
-  type VisualizerDynamicColor,
-  type AnalyzerResolution,
-  VISUALIZER_DEFAULTS,
-} from "../background";
+import { useVisualizerSettingsStore } from "../../stores/visualizer";
+import { useShallow } from "zustand/shallow";
 
 const Body = styled.div`
   display: flex;
@@ -46,45 +41,12 @@ function App() {
   const [preferences, setPreferences] = useState(defaultPreferences);
   const isMobile = useIsMobile();
 
-  // visualizer
-  const [visualizerType, setVisualizerType] = useState<VisualizerType>(
-    VISUALIZER_DEFAULTS.visualizerType
-  );
-  const [visualizerColorSource, setVisualizerColorSource] = useState<VisualizerColorSource>(
-    VISUALIZER_DEFAULTS.visualizerColorSource
-  );
-  const [visualizerSolidColor, setVisualizerSolidColor] = useState(
-    VISUALIZER_DEFAULTS.visualizerSolidColor
-  );
-  const [visualizerDynamicColor, setVisualizerDynamicColor] = useState<VisualizerDynamicColor>(
-    VISUALIZER_DEFAULTS.visualizerDynamicColor
-  );
-  const [visualizerLineThickness, setVisualizerLineThickness] = useState(
-    VISUALIZER_DEFAULTS.visualizerLineThickness
-  );
-  const [smoothVisualizer, setSmoothVisualizer] = useState(VISUALIZER_DEFAULTS.smoothVisualizer);
-  const [analyzerType, setAnalyzerType] = useState<Tone.AnalyserType>(
-    VISUALIZER_DEFAULTS.analyzerType
-  );
-  const [analyzerResolution, setAnalyzerResolution] = useState<AnalyzerResolution>(
-    VISUALIZER_DEFAULTS.analyzerResolution
-  );
-  const [frequencyGateX, setFrequencyGateX] = useState(VISUALIZER_DEFAULTS.frequencyGateX);
-  const [frequencyGateY, setFrequencyGateY] = useState(VISUALIZER_DEFAULTS.frequencyGateY);
-  const [frequencyGateTolerance, setFrequencyGateTolerance] = useState(
-    VISUALIZER_DEFAULTS.frequencyGateTolerance
-  );
-  const [frequencyGateInverted, setFrequencyGateInverted] = useState(
-    VISUALIZER_DEFAULTS.frequencyGateInverted
-  );
-  const [frequencyGateSustained, setFrequencyGateSustained] = useState(
-    VISUALIZER_DEFAULTS.frequencyGateSustained
-  );
-  const [viewGateVisual, setViewGateVisual] = useState(VISUALIZER_DEFAULTS.viewGateVisual);
-  const [visualizerSpeed, setVisualizerSpeed] = useState(VISUALIZER_DEFAULTS.visualizerSpeed);
-  const [hideMashupControls, setHideMashupControls] = useState(
-    VISUALIZER_DEFAULTS.hideMashupControls
-  );
+  const {saveAndLoadIndex: setVisualizerSettingsIndex, analyzerResolution, analyzerType} = useVisualizerSettingsStore(useShallow((state) => ({analyzerType: state.analyzerType, analyzerResolution: state.analyzerResolution, saveAndLoadIndex: state.saveAndLoadIndex})));
+
+  const [hideMashupControls, setHideMashupControls] = useState(false);
+
+  const analyzer = useMemo(() => new Tone.Analyser("fft", analyzerResolution), []);
+  const mashupTitles = useMemo(() => mashups.map((mashup) => mashup.mashedTrack.title), [mashups]);
 
   const fetchMashups = async () => {
     try {
@@ -103,8 +65,11 @@ function App() {
     fetchMashups();
   }, []);
 
-  const analyzer = useMemo(() => new Tone.Analyser("fft", analyzerResolution), []);
-  const mashupTitles = useMemo(() => mashups.map((mashup) => mashup.mashedTrack.title), [mashups]);
+  useEffect(() => {
+    setVisualizerSettingsIndex(mashupIndex);
+  }, [mashupIndex]);
+
+  
 
   useUpdate(() => {
     analyzer.type = analyzerType;
@@ -122,52 +87,9 @@ function App() {
         analyzer={analyzer}
         leftImageUrl={mashups[mashupIndex].track1.coverUrl}
         rightImageUrl={mashups[mashupIndex].track2.coverUrl}
-        visualizerType={visualizerType}
-        visualizerColorSource={visualizerColorSource}
-        visualizerSolidColor={visualizerSolidColor}
-        visualizerDynamicColor={visualizerDynamicColor}
-        visualizerLineThickness={visualizerLineThickness}
-        frequencyGateX={frequencyGateX}
-        frequencyGateY={frequencyGateY}
-        frequencyGateTolerance={frequencyGateTolerance}
-        frequencyGateInverted={frequencyGateInverted}
-        frequencyGateSustained={frequencyGateSustained}
-        viewGateVisual={viewGateVisual}
-        smoothVisualizer={smoothVisualizer}
-        visualizerSpeed={visualizerSpeed}
       />
       <Body>
         <NavBar
-          visualizerType={visualizerType}
-          setVisualizerType={setVisualizerType}
-          visualizerColorSource={visualizerColorSource}
-          setVisualizerColorSource={setVisualizerColorSource}
-          visualizerSolidColor={visualizerSolidColor}
-          setVisualizerSolidColor={setVisualizerSolidColor}
-          visualizerDynamicColor={visualizerDynamicColor}
-          setVisualizerDynamicColor={setVisualizerDynamicColor}
-          visualizerLineThickness={visualizerLineThickness}
-          setVisualizerLineThickness={setVisualizerLineThickness}
-          analyzerType={analyzerType}
-          setAnalyzerType={setAnalyzerType}
-          analyzerResolution={analyzerResolution}
-          setAnalyzerResolution={setAnalyzerResolution}
-          frequencyGateX={frequencyGateX}
-          setFrequencyGateX={setFrequencyGateX}
-          frequencyGateY={frequencyGateY}
-          setFrequencyGateY={setFrequencyGateY}
-          frequencyGateTolerance={frequencyGateTolerance}
-          setFrequencyGateTolerance={setFrequencyGateTolerance}
-          frequencyGateInverted={frequencyGateInverted}
-          setFrequencyGateInverted={setFrequencyGateInverted}
-          frequencyGateSustained={frequencyGateSustained}
-          setFrequencyGateSustained={setFrequencyGateSustained}
-          viewGateVisual={viewGateVisual}
-          setViewGateVisual={setViewGateVisual}
-          smoothVisualizer={smoothVisualizer}
-          setSmoothVisualizer={setSmoothVisualizer}
-          visualizerSpeed={visualizerSpeed}
-          setVisualizerSpeed={setVisualizerSpeed}
           setHideMashupControls={setHideMashupControls}
         />
         <Container $isMobile={isMobile} style={{visibility: hideMashupControls ? "hidden" : "visible"}}>
